requirements.txt
# ABOUTME: Python dependencies for the retail analytics dashboard project.
# ABOUTME: Includes data processing, database, web framework, and testing packages.

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# Web framework
streamlit>=1.28.0

# Visualization
plotly>=5.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
